Include the test-ludum.js for the minified tests.
This is instead of listing all the files individually. It makes sense to do that for
the unminified tests, but not so much for the minified ones. And this means
I only have to update tests/index.html whenever I add a new test file, I won't
have to update test/index-minified.html as well.