Don't throw away extra methods in StateMachine.addState
The addState method was only copying the update(), enter() and leave()
methods from kwargs, so there was no way to add the draw() method
expected by the game state machine (from gameloop.js).

Changed it so that if there is a kwargs object, we use it instead of just
copying bits of it over. This feels slightly less robust, but it fixes the bug.