Pass a StateMachine as first param to per-state functions
Changed the expected signature for the enter, update and leave functions
that you provide for each state:

enter()     --> enter(sm)
update(dt)  --> update(sm, dt)
leave()     --> leave(sm)

Where 'sm' is a reference to the StateMachine instance that the state belongs
to. This helps avoid the need to have a global variable for each state machine
instance, which is an improvement.