Pass a StateMachine as first param to draw functions
If the current state of the game state machine has a draw() function, the game
loop will automatically call it. Previously draw() was called with no
arguments. This commit makes the game loop pass a reference to the game state
machine instance as the first parameter of the draw function. This matches the
new behaviour of the other per-state functions, which also now take the state
machine as their first parameter.