Make every state machine keep track of time
Every state machine now has enterT and stateT attributes:
- enterT is the time it entered its current state.
- stateT is how long it's been in the current state for (as a convenience).

These are automatically updated by the start() and update() methods of the
state machine. They assume that ludum.updateCurrentTime() is being called once
per frame.

This change lead to a refactoring of the gameloop code. We no longer create a
state machine the we export; instead we require users to create and populate a
state machine which they pass in to the startGame method as a parameter. A nice
side effect is that the game code gets simpler from this too.