Try to get panning around the level working properly
The aim is to have the camera follow the player as far as possible, up to the
edge of the level, when they move outside the middle 50% of the viewport. It's
not working properly yet.

I've put in some code for drawing a background and some debugging info about
positions being displayed up in the top-left corner.