Player dies slowly and its game over...
When the players health reaches zero, we switch into the dying state (just like
for enemies). Once that's done, the player is dead and the "Game Over" screen
comes up. From there, pressing space or enter takes you back to the menu.

So now there's a loss condition, but not yet any victory condition.