Auto-start assets added after the loader has been started.
Previously, if you added a new asset to the loader after starting it, the asset
would never get loaded because the new asset wouldn't be started
automatically and there was no manual way to start it.

Now we have a flag which tracks whether the loader has been started. If it
has, any new assets you add to it get started automatically.